backend/pyproject.toml
[project]
name = "nous-center"
version = "0.1.0"
description = "AI media workstation - image/video/tts generation & multimodal understanding"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "celery[redis]>=5.4",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "httpx[socks]>=0.28",
    "websockets>=14.0",
    "modelscope>=1.34.0",
    "python-multipart>=0.0.22",
    "bcrypt>=4.0",
    "psutil>=6.0",
    "vllm>=0.19.0",
]

[tool.uv]
override-dependencies = [
    "transformers @ git+https://github.com/huggingface/transformers.git",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "ruff>=0.9",
    "aiosqlite>=0.22.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint.per-file-ignores]
# Late imports are intentional — either for model registration order,
# circular-import avoidance, or post-`create_app` exception handler wiring.
"src/api/main.py" = ["E402"]
"src/api/routes/workflows.py" = ["E402"]
"src/services/log_db.py" = ["E402"]
"src/workers/tts_engines/voxcpm2.py" = ["E402"]
"tests/conftest.py" = ["E402"]
"tests/test_errors.py" = ["E402"]
